Moving types Block and Transaction to respective folders
