+++ src/Blockcore/Consensus/Block/Block.cs
