+++ src/Blockcore/Consensus/Block/BlockHeader.cs
