+++ src/Blockcore/Consensus/Block/BlockSignature.cs
