+++ src/Blockcore/Consensus/Block/BlockStake.cs
