+++ src/Blockcore/Consensus/Block/ProvenBlockHeader.cs
