+++ src/Blockcore/Consensus/Transaction/PrecomputedTransactionData.cs
@@ -8,16 +8,19 @@
             this.HashSequence = Script.GetHashSequence(tx);
             this.HashPrevouts = Script.GetHashPrevouts(tx);
         }
+
         public uint256 HashPrevouts
         {
             get;
             set;
         }
+
         public uint256 HashSequence
         {
             get;
             set;
         }
+
         public uint256 HashOutputs
         {
             get;
