+++ src/Blockcore/Consensus/Transaction/Transaction.cs
