+++ src/Blockcore/Consensus/Transaction/TransactionSignature.cs
