+++ src/Blockcore/Interfaces/IBlockRepository.cs
