+++ src/Blockcore/NBitcoin/BitWriter.cs
@@ -8,6 +8,7 @@ namespace NBitcoin
     internal class BitReader
     {
         private BitArray array;
+
         public BitReader(byte[] data, int bitCount)
         {
             var writer = new BitWriter();
@@ -18,7 +19,7 @@ namespace NBitcoin
         public BitReader(BitArray array)
         {
             this.array = new BitArray(array.Length);
-            for(int i = 0; i < array.Length; i++)
+            for (int i = 0; i < array.Length; i++)
                 this.array.Set(i, array.Get(i));
         }
 
@@ -38,7 +39,7 @@ namespace NBitcoin
         public uint ReadUInt(int bitCount)
         {
             uint value = 0;
-            for(int i = 0; i < bitCount; i++)
+            for (int i = 0; i < bitCount; i++)
             {
                 uint v = Read() ? 1U : 0U;
                 value += (v << i);
@@ -57,7 +58,7 @@ namespace NBitcoin
         public BitArray ToBitArray()
         {
             var result = new BitArray(this.array.Length);
-            for(int i = 0; i < this.array.Length; i++)
+            for (int i = 0; i < this.array.Length; i++)
                 result.Set(i, this.array.Get(i));
             return result;
         }
@@ -76,11 +77,11 @@ namespace NBitcoin
 
         public bool Same(BitReader b)
         {
-            while(this.Position != this.Count && b.Position != b.Count)
+            while (this.Position != this.Count && b.Position != b.Count)
             {
                 bool valuea = Read();
                 bool valueb = b.Read();
-                if(valuea != valueb)
+                if (valuea != valueb)
                     return false;
             }
             return true;
@@ -89,9 +90,9 @@ namespace NBitcoin
         public override string ToString()
         {
             var builder = new StringBuilder(this.array.Length);
-            for(int i = 0; i < this.Count; i++)
+            for (int i = 0; i < this.Count; i++)
             {
-                if(i != 0 && i % 8 == 0)
+                if (i != 0 && i % 8 == 0)
                     builder.Append(' ');
                 builder.Append(this.array.Get(i) ? "1" : "0");
             }
@@ -102,6 +103,7 @@ namespace NBitcoin
     internal class BitWriter
     {
         private List<bool> values = new List<bool>();
+
         public int Count
         {
             get
@@ -109,6 +111,7 @@ namespace NBitcoin
                 return this.values.Count;
             }
         }
+
         public void Write(bool value)
         {
             this.values.Insert(this.Position, value);
@@ -140,11 +143,11 @@ namespace NBitcoin
         private static byte[] ToByteArray(BitArray bits)
         {
             int arrayLength = bits.Length / 8;
-            if(bits.Length % 8 != 0)
+            if (bits.Length % 8 != 0)
                 arrayLength++;
             var array = new byte[arrayLength];
 
-            for(int i = 0; i < bits.Length; i++)
+            for (int i = 0; i < bits.Length; i++)
             {
                 int b = i / 8;
                 int offset = i % 8;
@@ -152,7 +155,6 @@ namespace NBitcoin
             }
             return array;
         }
-
 
         public BitArray ToBitArray()
         {
@@ -165,14 +167,13 @@ namespace NBitcoin
             return Wordlist.ToIntegers(array);
         }
 
-
         private static byte[] SwapEndianBytes(byte[] bytes)
         {
             var output = new byte[bytes.Length];
-            for(int i = 0; i < output.Length; i++)
+            for (int i = 0; i < output.Length; i++)
             {
                 byte newByte = 0;
-                for(int ib = 0; ib < 8; ib++)
+                for (int ib = 0; ib < 8; ib++)
                 {
                     newByte += (byte)(((bytes[i] >> ib) & 1) << (7 - ib));
                 }
@@ -181,11 +182,9 @@ namespace NBitcoin
             return output;
         }
 
-
-
         public void Write(uint value, int bitCount)
         {
-            for(int i = 0; i < bitCount; i++)
+            for (int i = 0; i < bitCount; i++)
             {
                 Write((value & 1) == 1);
                 value = value >> 1;
@@ -193,6 +192,7 @@ namespace NBitcoin
         }
 
         private int _Position;
+
         public int Position
         {
             get
@@ -207,7 +207,7 @@ namespace NBitcoin
 
         internal void Write(BitReader reader, int bitCount)
         {
-            for(int i = 0; i < bitCount; i++)
+            for (int i = 0; i < bitCount; i++)
             {
                 Write(reader.Read());
             }
@@ -217,9 +217,10 @@ namespace NBitcoin
         {
             Write(bitArray, bitArray.Length);
         }
+
         public void Write(BitArray bitArray, int bitCount)
         {
-            for(int i = 0; i < bitCount; i++)
+            for (int i = 0; i < bitCount; i++)
             {
                 Write(bitArray.Get(i));
             }
@@ -238,14 +239,13 @@ namespace NBitcoin
         public override string ToString()
         {
             var builder = new StringBuilder(this.values.Count);
-            for(int i = 0; i < this.Count; i++)
+            for (int i = 0; i < this.Count; i++)
             {
-                if(i != 0 && i % 8 == 0)
+                if (i != 0 && i % 8 == 0)
                     builder.Append(' ');
                 builder.Append(this.values[i] ? "1" : "0");
             }
             return builder.ToString();
         }
     }
-
 }
