+++ src/Blockcore/NBitcoin/Repository/ITransactionRepository.cs
