+++ src/Blockcore/NBitcoin/Repository/InMemoryNoSqlRepository.cs
