+++ src/Blockcore/NBitcoin/Repository/NoSqlRepository.cs
