+++ src/Blockcore/NBitcoin/Repository/NoSqlTransactionRepository.cs
