+++ src/Blockcore/Utilities/ReaderWriterLock.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading;
 
-namespace NBitcoin
+namespace Blockcore.Utilities
 {
     /// <summary>
     /// Wraps ReaderWriterLockSlim with disposable interface so that
